tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/from_keyword.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - 'from' keyword in keyframe selector</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#keyframes">
   <meta name="flags" content="dom">
   <meta name="assert" content="In a keyframe selector, the keyword 'from' is
   equivalent to the value 0%">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes step-right {
         0% {
            left: 0px;
         }
         100% {
            left: 100px;
         }
         from {
            left: 100px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         position: relative;
      }
      .test {
         animation: step-right 1s;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         div.addEventListener('animationstart', function() {
            clearTimeout(timeout);
            var passed = getComputedStyle(div, null).left == '100px';
            test(function(){assert_true(passed);}, testName); done();
            log.data = passed ? 'PASS if the animation occurred.' : 'FAIL';
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The square below should immediately jump right, then after one second
   jump back to a left-aligned position.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
